Move "ast" submodule to toplevel
