+++ vunk-parser/src/lib.rs
@@ -2,7 +2,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-pub mod ast;
+pub mod expr;
+pub mod generic;
+pub mod literal;
+pub mod op;
+pub mod program;
 
 use vunk_lexer::Span;
