+++ vunk-parser/src/mod.rs
@@ -2,8 +2,3 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-pub mod expr;
-pub mod generic;
-pub mod literal;
-pub mod op;
-pub mod program;
