+++ vunk-parser/src/program.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::ast::expr::Expr;
+use crate::expr::Expr;
 
 #[derive(Debug)]
 #[cfg_attr(test, derive(PartialEq))]
